pyproject.toml
[project]
name = "repertoire-api"
version = "0.1.0"
description = "API for repertoire tasks"
authors = [
    {name = "Tomas Sheers"},
]
dependencies = [
    "sqlalchemy>=2.0.25",
    "pydantic>=2.6.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0.post1",
    "pydantic-settings>=2.1.0",
    "psycopg2>=2.9.9",
    "fast-depends>=2.3.1",
    "blinker>=1.8.2",
    "werkzeug>=3.0.3",
    "typer>=0.12.3",
    "pyjwt>=2.8.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
factory-boy = []
test = [
    "factory-boy>=3.3.0",
    "werkzeug>=3.0.1",
    "pytest>=8.0.0",
    "pytest-xdist>=3.5.0",
    "pytest-subtests>=0.11.0",
]
[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = [
    "E501",
    "F811",
    "F541",
    "F841"
]
